requirements.txt
amqp==2.5.1
asn1crypto==0.24.0
astroid==2.2.5
billiard==3.6.1.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
Django==2.2.4
djangorestframework==3.10.2
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-language==1.3.0
googleapis-common-protos==1.6.0
grpcio==1.23.0
idna==2.8
importlib-metadata==0.19
isort==4.3.21
kombu==4.6.4
lazy-object-proxy==1.4.2
mccabe==0.6.1
MeaningCloud-python==1.1.1
protobuf==3.9.1
psycopg2-binary==2.8.3
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycparser==2.19
pylint==2.3.1
pyOpenSSL==19.0.0
python-dotenv==0.10.3
pytz==2019.2
redis==3.3.8
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.3
vine==1.3.0
wrapt==1.11.2
zipp==0.5.2
